Move game types to a dedicated module
Add a dedicated module for game-related types and update imports
- New src/types/game.ts with TileType, CellCoord, TileAssignments
- New src/types/messages.ts referencing game and room types
- Shrink src/types/room.ts to Player and PlayerStatus only
- Update Board.tsx and useTileAssignment.ts to import from @/types/game
